requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-cors-headers==2.5.2
django-rest-framework==0.1.0
djangorestframework==3.9.0
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
psycopg2-binary==2.7.6.1
PyJWT==1.7.1
pytz==2018.7
whitenoise==4.0
django-registration==3.0.1
django-password-reset
python-dateutil
requests_cache
requests
google-cloud-logging
SQLAlchemy